+++ src/grandorgue/midi/files/GOMidiFileReader.h
@@ -1,6 +1,6 @@
 /*
  * Copyright 2006 Milan Digital Audio LLC
- * Copyright 2009-2023 GrandOrgue contributors (see AUTHORS)
+ * Copyright 2009-2026 GrandOrgue contributors (see AUTHORS)
  * License GPL-2.0 or later
  * (https://www.gnu.org/licenses/old-licenses/gpl-2.0.html).
  */
@@ -12,6 +12,8 @@
 #include <stdint.h>
 
 #include <wx/string.h>
+
+#include <vector>
 
 #include "GOBuffer.h"
 
@@ -44,6 +46,10 @@ public:
 
   bool Open(wxString filename);
   bool ReadEvent(GOMidiEvent &e);
+
+  /** Reads all remaining MIDI events, in time order. */
+  void ReadAllEvents(std::vector<GOMidiEvent> &outEvents);
+
   bool Close();
 };